Set "sane" default rotation settings
These should be the default rotation settings in Cultris II.

The change is made as @zDEFz had issues getting the default C2 settings
back and erroneously setting the S and Z pieces to rotation state to 1
and 3 respectively. This led to needing two 180 spins to get the pieces
to teleport down "diagonally", as opposed to when the defaults are set
so that S is 3 and Z is 1, in which case only one 180 is needed.