Changed the format of the configuration file.
My aim was to add a possibility to save the color of the pieces.
I found out that this is a bigger change, so I started out with this
and will create a discussion about the colors.
The configuration is easier to read and it's easier to add new data.