Add external projects to download and fetch dependencies, including curl
Off by default. Pass argument -DBUILD_THIRD_PARTY=ON to cmake to turn
on. At the moment, only works on Windows as Linux fails to build curl
due to a dependency on openssl. This superbuild is based on
https://github.com/Sarcasm/cmake-superbuild.

A note about curl: find_package(CURL) looks for curl include files and
libraries in specific directories (see find_path and find_library).
On Windows, set the cmake include and library path variables
CMAKE_INCLUDE_PATH and CMAKE_LIBRARY_PATH to the curl include and lib
directories respectively. On the command line, pass arguments
-DCMAKE_INCLUDE_PATH=<include dir> -DCMAKE_LIBRARY_PATH=<lib dir>.
The BUILD_THIRD_PARTY option handles this automatically.

Boost is not handled by the superbuild yet.