Changed to a much more conservative approach of using less thread
Using only 1 draw thread as multiple draw threads seemed to cause way too
much headache as neither SFML or TGUI is thread-safe.
The AI now also runs on 1 single thread. Possibly extend this as when
you have alot of them they play too slow and the CPU usage was not hitting
100%.
Also added an important feature where the updating and drawing of obsFields
will be skipped if there is no time for it, hence a higher priority is put
on updating and drawing the main playfield.