handleEvent, piecePreview, TextureAndSprite, ready/away, minor fixes
handleEvent for game-input is now handled in the polymorphic state class
Made a new derived TextureAndSprite widget for tgui that simply holds a
public sprite and texture. This was in turn used for the piecePreview in
GameOptions to remove a ugly hack and resolve the Mac issue with black boxes.
Ready and Away buttons and signaling should now work as intended again.
game.delayCheck is properly called during practice, no ai thread is
launched if there are no AI.

Updated to version 0.1.13 for next release!